Move selectAndPlay to player actions/reducer and call it start. Same for Queue
